+++ backend/public/css/app.css
@@ -10928,72 +10928,103 @@ a.text-dark:focus {
   }
 }
 
-html {
-  font-family: sans-serif;
-  line-height: 1.15;
+.top-wrapper {
+  background-color: #F6F6F6;
+  padding: 40px 0px;
+  height: 100vh;
 }
 
-body {
-  font-size: 16px;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Helvetica Neue, Hiragino Kaku Gothic ProN, "\30E1\30A4\30EA\30AA", meiryo, sans-serif;
-  font-weight: 400;
+.top-wrapper .articles-wrapper {
+  background-color: white;
+  margin: 0px auto;
+  padding: 20px;
 }
 
-nav {
-  background-color: #55C501 !important;
+.top-wrapper .articles-wrapper .article-box {
+  display: flex;
+  align-items: center;
+  padding: 20px;
+  border-bottom: solid 3px #F6F6F6;
 }
 
-nav #Navbar .navbar-nav .nav-item #post-link {
-  background-color: #367E00;
+.top-wrapper .articles-wrapper .article-box .article-box-left {
+  height: 40px;
+  width: 40px;
+  background-color: gray;
   border-radius: 5px;
-  padding-left: 1rem;
-  padding-right: 1rem;
 }
 
-nav #Navbar .navbar-nav .nav-item #register {
-  border: solid 2px white;
-  border-radius: 10px;
+.top-wrapper .articles-wrapper .article-box .article-box-right {
+  padding-left: 20px;
 }
 
-#login-wrapper {
-  background-color: #55C501;
+.top-wrapper .articles-wrapper .article-box .article-box-right .article-title {
+  font-weight: bold;
 }
 
-#login-wrapper .col-7 {
-  padding-left: 80px;
-  padding-top: 100px;
-  padding-bottom: 100px;
+.top-wrapper .articles-wrapper .article-box .article-box-right .article-date {
+  color: gray;
 }
 
-#login-wrapper .col-7 h1 {
-  font-size: 50px;
+.item-page-wrapper {
+  background-color: #f8f9fc;
 }
 
-#login-wrapper .col-5 form {
-  margin-top: 50px;
-  margin-right: 30px;
+.item-page-wrapper .item-wrapper {
+  background-color: white;
+  margin: 3rem 5rem;
+  padding: 3rem;
 }
 
-#login-wrapper .col-5 form table {
-  border-collapse: separate;
-  border-spacing: 0px 15px;
+.item-page-wrapper .item-wrapper .item-header .date {
+  color: gray;
 }
 
-#login-wrapper .col-5 form table tr th {
-  font-size: 10px;
+.item-page-wrapper .item-wrapper .item-title {
+  font-size: 2rem;
+  margin-top: 1.2rem;
+  font-weight: bold;
+}
+
+.item-page-wrapper .item-wrapper .item-tags .item-tag {
+  background-color: #dddddd;
+  display: inline-flex;
+  padding: 0.3rem;
+  color: #666;
+}
+
+.item-page-wrapper .item-wrapper .item-body {
+  margin-top: 2rem;
+}
+
+.item-page-wrapper .item-wrapper .item-body h1 {
+  font-size: 1.8rem;
+  font-weight: 700;
+  margin-bottom: 1rem;
+  padding-bottom: 0.2rem;
+  border-bottom: solid 1px #ddd;
+}
+
+.item-page-wrapper .item-wrapper .item-body h2 {
+  font-size: 1.5rem;
+  margin-bottom: 1rem;
+  padding-bottom: 0.2rem;
+  border-bottom: solid 1px #ddd;
+}
+
+.item-page-wrapper .item-wrapper .item-body h3 {
+  font-size: 1.2rem;
+  margin-bottom: 1rem;
+  padding-bottom: 0.2rem;
+  border-bottom: solid 1px #ddd;
+}
+
+.item-page-wrapper .item-wrapper .item-body pre code {
   color: white;
-  width: 80px;
-}
-
-#login-wrapper .col-5 form table tr td .form-control {
-  height: 45px;
-}
-
-#login-wrapper .col-5 form table tr td input[type=submit] {
-  width: 110px;
-  color: white;
-  background-color: #55C501;
-  border: solid 2px white;
+  padding: 1.5rem 1rem;
+  background-color: #364549;
+  width: 100%;
+  display: block;
 }
 
 .post-page-wrapper #title-input {
@@ -11087,102 +11118,71 @@ nav #Navbar .navbar-nav .nav-item #register {
   margin-right: 1rem !important;
 }
 
-.item-page-wrapper {
-  background-color: #f8f9fc;
+nav {
+  background-color: #55C501 !important;
 }
 
-.item-page-wrapper .item-wrapper {
-  background-color: white;
-  margin: 3rem 5rem;
-  padding: 3rem;
-}
-
-.item-page-wrapper .item-wrapper .item-header .date {
-  color: gray;
-}
-
-.item-page-wrapper .item-wrapper .item-title {
-  font-size: 2rem;
-  margin-top: 1.2rem;
-  font-weight: bold;
-}
-
-.item-page-wrapper .item-wrapper .item-tags .item-tag {
-  background-color: #dddddd;
-  display: inline-flex;
-  padding: 0.3rem;
-  color: #666;
-}
-
-.item-page-wrapper .item-wrapper .item-body {
-  margin-top: 2rem;
-}
-
-.item-page-wrapper .item-wrapper .item-body h1 {
-  font-size: 1.8rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-  padding-bottom: 0.2rem;
-  border-bottom: solid 1px #ddd;
-}
-
-.item-page-wrapper .item-wrapper .item-body h2 {
-  font-size: 1.5rem;
-  margin-bottom: 1rem;
-  padding-bottom: 0.2rem;
-  border-bottom: solid 1px #ddd;
-}
-
-.item-page-wrapper .item-wrapper .item-body h3 {
-  font-size: 1.2rem;
-  margin-bottom: 1rem;
-  padding-bottom: 0.2rem;
-  border-bottom: solid 1px #ddd;
-}
-
-.item-page-wrapper .item-wrapper .item-body pre code {
-  color: white;
-  padding: 1.5rem 1rem;
-  background-color: #364549;
-  width: 100%;
-  display: block;
-}
-
-.top-wrapper {
-  background-color: #F6F6F6;
-  padding: 40px 0px;
-  height: 100vh;
-}
-
-.top-wrapper .articles-wrapper {
-  background-color: white;
-  margin: 0px auto;
-  padding: 20px;
-}
-
-.top-wrapper .articles-wrapper .article-box {
-  display: flex;
-  align-items: center;
-  padding: 20px;
-  border-bottom: solid 3px #F6F6F6;
-}
-
-.top-wrapper .articles-wrapper .article-box .article-box-left {
-  height: 40px;
-  width: 40px;
-  background-color: gray;
+nav #Navbar .navbar-nav .nav-item #post-link {
+  background-color: #367E00;
   border-radius: 5px;
+  padding-left: 1rem;
+  padding-right: 1rem;
 }
 
-.top-wrapper .articles-wrapper .article-box .article-box-right {
-  padding-left: 20px;
+nav #Navbar .navbar-nav .nav-item #register {
+  border: solid 2px white;
+  border-radius: 10px;
 }
 
-.top-wrapper .articles-wrapper .article-box .article-box-right .article-title {
-  font-weight: bold;
+#login-wrapper {
+  background-color: #55C501;
 }
 
-.top-wrapper .articles-wrapper .article-box .article-box-right .article-date {
-  color: gray;
+#login-wrapper .col-7 {
+  padding-left: 80px;
+  padding-top: 100px;
+  padding-bottom: 100px;
+}
+
+#login-wrapper .col-7 h1 {
+  font-size: 50px;
+}
+
+#login-wrapper .col-5 form {
+  margin-top: 50px;
+  margin-right: 30px;
+}
+
+#login-wrapper .col-5 form table {
+  border-collapse: separate;
+  border-spacing: 0px 15px;
+}
+
+#login-wrapper .col-5 form table tr th {
+  font-size: 10px;
+  color: white;
+  width: 80px;
+}
+
+#login-wrapper .col-5 form table tr td .form-control {
+  height: 45px;
+}
+
+#login-wrapper .col-5 form table tr td input[type=submit] {
+  width: 110px;
+  color: white;
+  background-color: #55C501;
+  border: solid 2px white;
+}
+
+html {
+  font-family: sans-serif;
+  line-height: 1.15;
+}
+
+body {
+  font-size: 16px;
+  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Helvetica Neue, Hiragino Kaku Gothic ProN, "\30E1\30A4\30EA\30AA", meiryo, sans-serif;
+  font-weight: 400;
 }
