+++ backend/public/js/app.js
@@ -52226,13 +52226,9 @@ module.exports = function(module) {
 /*!*****************************!*\
   !*** ./resources/js/app.js ***!
   \*****************************/
-/*! no exports provided */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony import */ var marked__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! marked */ "./node_modules/marked/lib/marked.js");
-/* harmony import */ var marked__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(marked__WEBPACK_IMPORTED_MODULE_0__);
 /**
  * First we will load all of this project's JavaScript dependencies which
  * includes Vue and other libraries. It is a great starting point when
@@ -52240,29 +52236,8 @@ __webpack_require__.r(__webpack_exports__);
  */
 __webpack_require__(/*! ./bootstrap */ "./resources/js/bootstrap.js");
 
+__webpack_require__(/*! ./parts/markdown */ "./resources/js/parts/markdown.js");
 
-$(function () {
-  marked__WEBPACK_IMPORTED_MODULE_0___default.a.setOptions({
-    langPrefix: '',
-    breaks: true,
-    sanitize: true
-  });
-  $('#markdown_editor_textarea').keyup(function () {
-    var html = marked__WEBPACK_IMPORTED_MODULE_0___default()(getHtml($(this).val()));
-    $('#markdown_preview').html(html);
-  }); // 個別の記事画面のマークダウンをHTMLに変換する
-
-  var target = $('.item-body');
-  var convertedHtml = marked__WEBPACK_IMPORTED_MODULE_0___default()(getHtml(target.html()));
-  target.html(convertedHtml); // 比較演算子が $lt; 等になるので置換
-
-  function getHtml(html) {
-    html = html.replace(/&lt;/g, '<');
-    html = html.replace(/&gt;/g, '>');
-    html = html.replace(/&amp;/g, '&');
-    return html;
-  }
-});
 window.Vue = __webpack_require__(/*! vue */ "./node_modules/vue/dist/vue.common.js");
 /**
  * The following block of code may be used to automatically register your
@@ -52398,6 +52373,43 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_ExampleComponent_vue_vue_type_template_id_299e239e___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
 
 
+
+/***/ }),
+
+/***/ "./resources/js/parts/markdown.js":
+/*!****************************************!*\
+  !*** ./resources/js/parts/markdown.js ***!
+  \****************************************/
+/*! no exports provided */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var marked__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! marked */ "./node_modules/marked/lib/marked.js");
+/* harmony import */ var marked__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(marked__WEBPACK_IMPORTED_MODULE_0__);
+
+$(function () {
+  marked__WEBPACK_IMPORTED_MODULE_0___default.a.setOptions({
+    langPrefix: '',
+    breaks: true,
+    sanitize: true
+  });
+  $('#markdown_editor_textarea').keyup(function () {
+    var html = marked__WEBPACK_IMPORTED_MODULE_0___default()(getHtml($(this).val()));
+    $('#markdown_preview').html(html);
+  }); // 個別の記事画面のマークダウンをHTMLに変換する
+
+  var target = $('.item-body');
+  var convertedHtml = marked__WEBPACK_IMPORTED_MODULE_0___default()(getHtml(target.html()));
+  target.html(convertedHtml); // 比較演算子が $lt; 等になるので置換
+
+  function getHtml(html) {
+    html = html.replace(/&lt;/g, '<');
+    html = html.replace(/&gt;/g, '>');
+    html = html.replace(/&amp;/g, '&');
+    return html;
+  }
+});
 
 /***/ }),
