+++ backend/resources/sass/app.scss
@@ -8,6 +8,22 @@
 @import '~bootstrap/scss/bootstrap';
 
 
+// top
+@import './top';
+
+// ????
+@import './item-page';
+
+// ??????
+@import './post-page';
+
+// header
+@import './parts/header';
+
+// login
+@import './parts/login';
+
+
 html {
     font-family: sans-serif;
     line-height: 1.15;
@@ -17,264 +33,4 @@ body {
     font-size: 16px;
     font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Helvetica Neue, Hiragino Kaku Gothic ProN, "メイリオ", meiryo, sans-serif;
     font-weight: 400;
-}
-
-nav {
-    background-color: #55C501 !important;
-
-    #Navbar {
-        .navbar-nav {
-            .nav-item {
-                #post-link {
-                    background-color: #367E00;
-                    border-radius: 5px;
-                    padding-left: 1rem;
-                    padding-right: 1rem;
-                }
-                #register {
-                    border: solid 2px white;
-                    border-radius: 10px;
-                }
-            }
-        }
-    }
-}
-
-#login-wrapper {
-    background-color: #55C501;
-
-    .col-7 {
-        padding-left: 80px;
-        padding-top: 100px;
-        padding-bottom: 100px;
-
-        h1 {
-            font-size: 50px;
-        }
-    }
-
-    .col-5 {
-        form {
-            margin-top: 50px;
-            margin-right: 30px;
-
-            table {
-                border-collapse: separate;
-                border-spacing: 0px 15px;
-
-                tr {
-                    th {
-                        font-size: 10px;
-                        color: white;
-                        width: 80px;
-                    }
-
-                    td {
-                        .form-control {
-                            height: 45px;
-                        }
-
-                        input[type="submit"] {
-                            width: 110px;
-                            color: white;
-                            background-color: #55C501;
-                            border: solid 2px white;
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-// ??????
-.post-page-wrapper {
-    #title-input {
-        height: 3rem;
-
-        &::placeholder {
-            font-size: 1.3rem;
-        }
-    }
-
-    .row {
-        justify-content: center;
-
-        .col-6 {
-            padding: 0rem;
-            flex: none;
-            width: 48%;
-
-            #markdown_editor_textarea {
-                padding: 1rem;
-                height: 29rem;
-                width: 100%;
-            }
-
-            #markdown_preview {
-                padding: 1rem;
-                height: 29rem;
-                border: solid 1px gray;
-                width: 100%;
-                background-color: white;
-                border: 1px solid #ced4da;
-                border-radius: 0.25rem;
-                color: #495057;
-
-                h1 {
-                    font-size: 2rem;
-                    font-weight: 700;
-                    margin-bottom: 1rem;
-                    padding-bottom: 0.2rem;
-                    border-bottom: solid 1px #ddd;
-                }
-
-                h2 {
-                    font-size: 1.8rem;
-                    margin-bottom: 1rem;
-                    padding-bottom: 0.2rem;
-                    border-bottom: solid 1px #ddd;
-                }
-
-                h3 {
-                    font-size: 1.6rem;
-                    margin-bottom: 1rem;
-                    padding-bottom: 0.2rem;
-                    border-bottom: solid 1px #ddd;
-                }
-
-                pre {
-                    code {
-                        color: white;
-                        padding: 1.5rem 1rem;
-                        background-color: #364549;
-                        width: 100%;
-                        display: block;
-                    }
-                }
-            }
-        }
-    }
-
-    .post-page-footer {
-        display: flex;
-        justify-content: flex-end;
-
-        .post-button {
-            background-color: #55C501;
-            color: white;
-            border-radius: 0.25rem;
-            width: 7rem;
-            height: 2rem;
-            margin-right: 1rem !important;
-        }
-    }
-}
-
-// ????
-.item-page-wrapper {
-    background-color: #f8f9fc;
-
-    .item-wrapper {
-        background-color: white;
-        margin: 3rem 5rem;
-        padding: 3rem;
-
-        .item-header {
-            .date {
-                color: gray;
-            }
-        }
-
-        .item-title {
-            font-size: 2rem;
-            margin-top: 1.2rem;
-            font-weight: bold;
-        }
-
-        .item-tags {
-            .item-tag {
-                background-color: rgb(221, 221, 221);
-                display: inline-flex;
-                padding: 0.3rem;
-                color: #666;
-            }
-        }
-
-        .item-body {
-            margin-top: 2rem;
-
-            h1 {
-                font-size: 1.8rem;
-                font-weight: 700;
-                margin-bottom: 1rem;
-                padding-bottom: 0.2rem;
-                border-bottom: solid 1px #ddd;
-            }
-
-            h2 {
-                font-size: 1.5rem;
-                margin-bottom: 1rem;
-                padding-bottom: 0.2rem;
-                border-bottom: solid 1px #ddd;
-            }
-
-            h3 {
-                font-size: 1.2rem;
-                margin-bottom: 1rem;
-                padding-bottom: 0.2rem;
-                border-bottom: solid 1px #ddd;
-            }
-
-            pre {
-                code {
-                    color: white;
-                    padding: 1.5rem 1rem;
-                    background-color: #364549;
-                    width: 100%;
-                    display: block;
-                }
-            }
-        }
-    }
-}
-
-
-// ?????????
-.top-wrapper {
-    background-color: #F6F6F6;
-    padding: 40px 0px;
-    height: 100vh;
-
-    .articles-wrapper {
-        background-color: white;
-        margin: 0px auto;
-        padding: 20px;
-
-        .article-box {
-            display: flex;
-            align-items: center;
-            padding: 20px;
-            border-bottom: solid 3px #F6F6F6;
-
-            .article-box-left {
-                height: 40px;
-                width: 40px;
-                background-color: gray;
-                border-radius: 5px;
-            }
-
-            .article-box-right {
-                padding-left: 20px;
-
-                .article-title {
-                    font-weight: bold;
-                }
-
-                .article-date {
-                    color: gray;
-                }
-            }
-        }
-    }
 }
