+++ backend/resources/sass/parts/header.scss
@@ -1,0 +1,21 @@
+nav {
+    background-color: #55C501 !important;
+
+    #Navbar {
+        .navbar-nav {
+            .nav-item {
+                #post-link {
+                    background-color: #367E00;
+                    border-radius: 5px;
+                    padding-left: 1rem;
+                    padding-right: 1rem;
+                }
+
+                #register {
+                    border: solid 2px white;
+                    border-radius: 10px;
+                }
+            }
+        }
+    }
+}
