+++ backend/resources/sass/parts/login.scss
@@ -1,0 +1,46 @@
+#login-wrapper {
+    background-color: #55C501;
+
+    .col-7 {
+        padding-left: 80px;
+        padding-top: 100px;
+        padding-bottom: 100px;
+
+        h1 {
+            font-size: 50px;
+        }
+    }
+
+    .col-5 {
+        form {
+            margin-top: 50px;
+            margin-right: 30px;
+
+            table {
+                border-collapse: separate;
+                border-spacing: 0px 15px;
+
+                tr {
+                    th {
+                        font-size: 10px;
+                        color: white;
+                        width: 80px;
+                    }
+
+                    td {
+                        .form-control {
+                            height: 45px;
+                        }
+
+                        input[type="submit"] {
+                            width: 110px;
+                            color: white;
+                            background-color: #55C501;
+                            border: solid 2px white;
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
