+++ backend/resources/sass/post-page.scss
@@ -1,0 +1,82 @@
+.post-page-wrapper {
+    #title-input {
+        height: 3rem;
+
+        &::placeholder {
+            font-size: 1.3rem;
+        }
+    }
+
+    .row {
+        justify-content: center;
+
+        .col-6 {
+            padding: 0rem;
+            flex: none;
+            width: 48%;
+
+            #markdown_editor_textarea {
+                padding: 1rem;
+                height: 29rem;
+                width: 100%;
+            }
+
+            #markdown_preview {
+                padding: 1rem;
+                height: 29rem;
+                border: solid 1px gray;
+                width: 100%;
+                background-color: white;
+                border: 1px solid #ced4da;
+                border-radius: 0.25rem;
+                color: #495057;
+
+                h1 {
+                    font-size: 2rem;
+                    font-weight: 700;
+                    margin-bottom: 1rem;
+                    padding-bottom: 0.2rem;
+                    border-bottom: solid 1px #ddd;
+                }
+
+                h2 {
+                    font-size: 1.8rem;
+                    margin-bottom: 1rem;
+                    padding-bottom: 0.2rem;
+                    border-bottom: solid 1px #ddd;
+                }
+
+                h3 {
+                    font-size: 1.6rem;
+                    margin-bottom: 1rem;
+                    padding-bottom: 0.2rem;
+                    border-bottom: solid 1px #ddd;
+                }
+
+                pre {
+                    code {
+                        color: white;
+                        padding: 1.5rem 1rem;
+                        background-color: #364549;
+                        width: 100%;
+                        display: block;
+                    }
+                }
+            }
+        }
+    }
+
+    .post-page-footer {
+        display: flex;
+        justify-content: flex-end;
+
+        .post-button {
+            background-color: #55C501;
+            color: white;
+            border-radius: 0.25rem;
+            width: 7rem;
+            height: 2rem;
+            margin-right: 1rem !important;
+        }
+    }
+}
